pyproject.toml
[project]
name = "loretta"
version = "0.1.0"
description = "Ein vielseitiger Discord-Bot"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "aiosqlite>=0.21.0",
    "discord-py[speed,voice]>=2.5.2",
    "feedparser>=6.0.11",
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "pandas>=2.3.1",
    "psutil>=7.0.0",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
    "discord-py[speed,voice]>=2.5.2",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.7",
]
[project.scripts]
loretta = "bot.main:main"

[tool.ruff]
lint.extend-select = ["I"]
